add the base Runner; add params-related task helper methods
This implements the basic Runner for running tasks.  This is a
base for building more specific runners for more specific cases:
running in tests, running dry-runs, etc.  The runner stores the
params for each task it runs.  This adds the params-related
runner logic and also adds two params-related helper methods
to the task.

The params is a Hash that errors if unknown keys are requested.
The goal here is to be noisy and alert the developer if they are
asking for a param that hasn't been set.  The params are built
from any given when the runner is created.  Eventually, the
runners will be built using a global config's params.

The task now has a `params` helper method for accessing the
params.  Task's can also be built using any given custom params
but fall back to the runner's params.  The task also now has a
`set_param` method that allows a task to set runner-level params.
This allows you to create tasks that just configure params on
the runner that change the behavior of other tasks.