stringify runner params
This ensures that no matter how the param key was set, you access
it commonly with string keys.

Note: to commonize this properly, I needed the runner to implement
the stringify function and have all runner params writing go
through the runner (ie the `set_param` method).  This should have
been done originally as all meaningful logic should be controlled
and commonized by the runner but was missed in the previous
effort.