setup the runner; break out runner API for running sub-tasks
This sets up the 4 different runners Dk will use so that new
features can be added with all runner concerns covered.  This
also breaks out a new `run_task` method on the base Runner that
is used for running sub-tasks (tasks from other tasks).  This is
needed to distinguish whether a task run is for a top level task
or a sub-task b/c the runner's behavior changes depending on this
(for example the test runner doesn't run sub-tasks).

The main DkRunner and the DryRunner simply subclass Runner for now.
THe main DkRunner will be used for running live tasks while the
DryRunner will be used for just logging task activity but not
actually running and task cmds.

The TestRunner will run tasks in tests.  This runner disabled
running sub-tasks but captures anything it runs (tasks or cmds in
the future).

The TreeRunner subclasses DryRunner b/c it too will run all the
sub-tasks with disabled cmds.  However, like the TestRunner, it
captures all sub-tasks run in a tree and will display this tree
in the future.  It will also disable logging in the future.