add task test helpers for building test runners and test tasks
These are intended to help developers with building runners and
tasks in tests.  The idea is they shouldn't have to know the
relationship of the runners and tasks to test their tasks.  They
shouldn't be building runners/tasks manually b/c their are things
that need to "line up".

I'm going for a similar api to the test helpers in our "handler
pattern" gems (ie Deas, Sanford and Qs).  That means altering the
signature of the test runner's run method to only take optional
params.  This means you configure a task class on the test runner
and calling run builds a task of that class and runs it.  I also
added a `task` method on the test runner to match the `handler`
method on the handler pattern test runners.

Note: I switched to using the test helpers on the task tests to
further validate they work as intended.  This was done in the
handler pattern gems too.