`halt` task helper method
This method halts execution on *only* the current task being run.
The goal is to give a helper to say "I'm done with this task" but
keep running through all the other tasks that are supposed to run.

Note: it isn't recommended you use this for error handling or in
error scenarios where you want to stop execution of not only the
current task but any subsequent tasks.  In this scenario, simply
raise an exception.

Note also: this reworks and cleans up some runner related logic
and tests to better match the order of the task helper methods
and to cleanup some things that were missed in previous efforts.