log task helper methods
This adds some helper methods to tasks for logging output.  Info
logging is for general task feedback to the user.  Debug logging
is intended for verbose feeback to the user.  Error logging is
for letting the user know an error has occurred.

This logging methods all call to the runner's logger.  This allows
different runners to supply different loggers to alter the logging
behavior based on their design needs.  For now, the base runner
defaults the logger to a null logger that does nothing.

In the future, the test runner and tree runner will continue just
using the default null logger.  These loggers should never show
any output to the user as they are for internal use only (ie you
don't want logs outputting while running a test suite or while
building the task tree for display).

In the future, the live/dry runners will configure a Logsly logger
on stdout. This logger will force the task feedback to stdout while
allowing the user to configure the stdout logger details for style,
etc.  There will also be an option to configure file logging.  The
dry runner will have the same logger as the live logger since it
is designed to imitate live logging as closely as possible.