add local cmd handling API on tasks and runners
This adds the API for running local cmds on tasks.  There are two
methods: `cmd` and `cmd!` for running local cmds.  Both run a
given cmd string - the bang version raises an exception if the
cmd is not successful.  Both return the local cmd output so you
can get data about the cmd (whether it succeeded or not, its
stdout, etc).

This also sets up the handling of local cmds in the runners.  All
local cmds are logged using the runner's logger.  The test runner
always builds cmd spies instead of live cmds so that cmds are
never run in tests.  The test runner also has a cmd stubbing API
for adding in custom cmd behavior and cmd spying to task tests.
Finally, the test runner captures cmd spies that are run in its
`runs`.  Use this to test that a task runs a local command.