local cmd and spy objects
These objects will be used to track and run local commands in
tasks.  In a coming PR, tasks will get an API for running local
cmds in their tasks.  That API will build these objects and run
them.  A spy is added b/c the test runner will only want to
create spies from its API.  Also the test runner will need an
API for users to stub in custom cmd spies to immitate task
behavior in tests.

The local cmd and spy both compose Scmd cmds and spies.  They both
provide a subset of Scmd's API designed for Dk use.  Both also
have the concept of "output lines".  This is the captured output
for the command in a structured format that will be used for
logging cmd output in a standard way.

Note: the coming remote cmd objects will have a similar, limited
API that matches the local cmd.  The main difference will be that
you can access the stdout/stderr of local cmds but will not be
able to in remote cmds.  This is b/c the value could be diffrent
across the number of hosts the remote cmd is run on.