rework test runner test task API
This is prep for adding local cmd handling and stubbing APIs to
the test runner.  These were noticed while planning for local
cmds and how the test runner will interact with them.

First, this forces that the test runner only ever runs its task.
The goal here is that if you get a task instance from a test
runner and the run the test runner, that specific task will be
the task that is run.  This keeps everything in line and
consistent for testing purposes.

Second, this reworks the test runner tests to prep for capturing
local cmd runs.  This makes the getting the test task state
easier and will allow for testing the test runner cmd stubbing
API in a coming PR.