remote cmd and spy objects
These objects will be used to track and run remote commands in
tasks. In a coming PR, tasks will get an API for running remote
cmds in their tasks. That API will build these objects and run
them. A spy is added b/c the test runner will only want to
create spies from its API. Also the test runner will need an
API for users to stub in custom cmd spies to immitate cmd
behavior in task tests.

The remote cmd and spy both compose local cmds and spies. A local
cmd/spy is built for each host specified.  Both the cmd and spy
provide a similar API to local cmds/spies, however that API is
applied to the set of local cmds. For example, when run, the cmd
runs all of its local cmds at the same time.  The cmd is only
successful if all of its local cmds are successful.  The cmd's
output lines are the concatenation of all of its local cmd's
output lines.  One difference with local cmds though: you cannot
access the stdout/stderr of remote cmds as this data could be
different across each host.

Note: I chose to expose the local cmd's Scmd as a reader.  This
is so I could access internal Scmd things for remote cmds usage.
Specifically the `start` and `wait` API needed for the run
logic.  I didn't want to demeter this in the local cmd's api
b/c it isn't designed to have this api (this is a demeter
violation though).  I also figured it might be nice to have
access to the Scmd in your tasks (in rare special cases like
getting the PID or something.

In a coming PR, I'll add in the ssh parts of the cmds to make
them actually "remote".  I'll also add in the runner/task
handling API (similar to the one for local cmds).

Note: the remote base cmd takes a local cmd/spy *class* and
initializes an instance for each host.  This is different than
the local cmd/spy that taskes an already initialized instance.
In a coming PR, I'm going to rework the local cmds to have
similar handling, if nothing else to be consistent with the
remote cmds.

Note also: the `cmd_str` and `to_s` are different than the
local cmds.  The remote cmd tracks this on its own instead
of demetering the local cmd.  This is b/c there isn't an
obvious choice of which local cmd to demeter and we might
as well track it manually b/c it is trivial and it is more
decoupled from the local cmd.  Also, the local cmd will
have all the "ssh markup" in it which I don't want to
display on remote cmds.  I'm going to update the local cmd
to match and be consistent.