local cmd modifications based on remote cmd needs/work
This is a set of changes to the local cmd objects based on R&D
and implementation work on remote cmds.  There are a couple of
small things here (however, the main goal of this is to keep the
implementations as similar as possible):

* have the base cmd take the local cmd/spy *class* and create an
  instance.  This has no specific value to local cmds, but remote
  cmds needed this API so I'm updating local cmds to match.
* track the cmd str and to_s manually (instead of demetering from
  the Scmd).  Again, no specific value here, but the remote cmd
  needed this so I want the local to match.  Also, this decouples
  slightly from the Scmd.

Finally, this removes the `run!` method from local cmds.  This
issue is that we were relying on Scmd to error in this case.  This
means the error was previously an Scmd error with a backtrace
buried in the local cmd internals.  A better implementation is to
handle this logic in the task (where the `cmd!` api is actually
needed).  This switches to just implementing/calling `run` on the
local cmd and runner.  For the task's `cmd!` method, the run meth
is called and then if the cmd was not successful, a custom task
exception is raised with the caller of the `cmd!` method as the
backtrace.  This results in better errors for the users.  Also,
this is handy b/c it lines up with the remote cmd api/handling.
See PR 13 for reference.