add remote cmd handling API on tasks and runners
This adds the API for running remote cmds on tasks.  There are two
methods: `ssh` and `ssh!` for running remote cmds.  Both run a
given cmd string - the bang version raises an exception if the
cmd is not successful.  Both return the remote cmd so you can
access its data (whether it succeeded or not, etc).

This also sets up the handling of remote cmds in the runners.  All
remote cmds are logged using the runner's logger.  The test runner
always builds cmd spies instead of live cmds so that cmds are
never run in tests.  The test runner also has a cmd stubbing API
for adding in custom cmd behavior and cmd spying to task tests.
Finally, the test runner captures cmd spies that are run in its
`runs`.  Use this to test that a task runs a remote command.

Note: I chose to name the task/runner method `ssh` (instead of the
more obvious `remote_cmd`) b/c it has the same number of chars as
`cmd` and because it more clearly communicates what is happening
under the covers.  I still like the "remote cmd" name for the
object b/c it more formally models the *concept* and is the ying
to the local cmd yang.  However the shorter name plus the explicit
intention revealing nature of the method name in the context of a
task wins out.  The runner is the naming cross-over point: it
shows the calling `ssh` means you are building and running a
remote cmd.

Note also: I chose to add `stdout=` and `stderr=` writers to the
remote cmd spy.  I figure it can't hurt (since the implementation
matches `exitstatus=`) and it allows creating spis that produce
output lines (which is nice for testing).