add ssh "markup" to RemoteCmd LocalCmd strings
This is needed to actually run the cmds on each host remotely.
This adds a new `:ssh_args` option for specifying custom SSH
cmd args.  It handles escaping the cmds appropriately and running
each cmd in a shell which ensures the full user profile is loaded.

Note: this updates the LocalCmd initializer but this has no effect
on the behavior.  The update is just to better match the new
RemoteCmd initializer.