add a ConfigRunner and have the dk/dry runners subclass it
The idea is that these runners (including the TreeRunner as it
subclasses the DryRunner) need to set their runner args from a
given configuration's attrs.  This will allow the user to config
settings for Dk runs and have those settings be honored.

Note: the reason this does not apply to the test runner is that
its logic is run in isolation:

* any expected config param values should be manually specified
  when building the test runner
* no callback config settings are needed as callbacks are ignored
* no ssh config settings are needed as all ssh stuff is spied
* no log config settings are needed since there is no logging
* no task config settings are needed since you are testing an
  individual task

All the Config settings should be unit tested separately using
Dk's config and requiring in the config file like the CLI would.

In future efforts, I'll add the relevant config settings and
have the runners pull them from the config as needed.