default the Remote::CmdSpy hosts if none are given
This is needed b/c if there are no hosts specified, the spy
methods will fail as they assume a `@first_local_cmd_spy` and
that is `nil` if there are no hosts.

This scenario will come up on ssh cmds that don't specify a custom
`:hosts` value at cmd-time (ie they use the ssh hosts configured
for the task).  This should have been accounted for back in PR 15
but was missed.