add `set_param` config method and use config params in config runners
This allows the user to configure a set of global, default param
values that will be available to all tasks being run.  The params
are normalized in the same manner the runner does to help prevent
any param name conflicts when they get applied on the runner.

This is part of setting up the config's API/DSL.