add a task callback API to the Config
This are just helper methods for setting the callbacks on the
subject task classes directly.  The one advantage is it reads
more like a DSL and the execution can be delayed by running in
a `configure` proc.  This allows you to configure callbacks
at the top of the callback task's file for readability and
reference the callback task before the constant has been
defined:

```ruby
require 'dk'
require 'my_task'

Dk.configure do
  before MyTask, MyOtherTask, :some => 'param'
end

class MyOtherTask
  include Dk::Task

end
```

Otherwise this is just a thin helper method and you can use it
or not.