add ssh dsl methods to the Config
This allows the user to configure named ssh host groups to use
when configuring ssh hosts for tasks.  This also allows the user
to configure global default ssh args to use in ssh cmds in tasks.

In a future effort, I'll update the tasks ssh usage to honor these
settings.