commonize the `set_param` logic between the Config and Runner
This is a cleanup now that both of these classes provide this API.
This gets the implementations identical b/c that is the intent.
This was noticed while working on common ssh opts between the
Config and Runner.  That API needed to be common and then I
noticed this one does too.  This should have been done in PR 22
but was missed.