add host-specific ssh args to Config DSL; honor when making ssh cmds
This allows users to set ssh args that are only needed/applicable
to specific hosts.  This could be things like one host uses a
custom SSH port while the rest do not, etc.

This adds a new DSL method, `host_ssh_args` to the Config DSL.
Given a host name and the args, it stores them off.  That data is
then sent to the runner who passes it on to the Remote::Cmd.  The
remote cmd inserts the args after the global args on the local
cmd for the given host.