build a Logsly logger from a config and use on the live/dry runners
This configured logsly logger will handle logging output to stdout
and optionally to a file when running tasks live in in dry-run
mode.

This adds a few config DSL methods to help control the logger. Use
`log_pattern` to change the pattern when logging to stdout. By
default only the log message will be output.  All stdout logging
will be at the INFO level unless in verbose mode then logging will
be at the DEBUG level.  Use `log_file` to turn on logging to a
file by setting a log file path relative to the PWD.  This is
optional.  All file logging will be at the DEBUG level, regardless
of whether in verbose mode or not.  The idea is that you can set
a log file to ensure you can always access the verbose details of
a task run while just using the non-verbose default on stdout.
Finally, use the `log_file_pattern` to change the pattern when
file logging.  By default the date, level and message will be
output.

This updates the ConfigRunner (used by the live, dry and tree
runners) to super with the config's logger unless a custom logger
has been given.  The TreeRunner now passes a NullLogger to turn
off logger (per its requirements).