add first-pass README docs for the config DSL
This is part of documenting Dk and its usage.  This covers the
Config and its DSL.  This also includes a number of tweaks to the
previous docs to generally improve them and make them more
readable.

In addition, I updated the `ssh_hosts` api to take a variable list
of values instead of forcing the values to be passed in as a
single array.  As I was adding example usage docs I found it was
a better way to configure long lists of servers.  This is also
backwards compatible with the previous single array API.