basic CLI with runner flags and list tasks options
This is the basic CLI for running callbable tasks.  You pass it
a list of task names and it runs each named task.  You can pass
the `--dry-run` option to use the DryRunner and the `--tree` opt
to use the TreeRunner.  Pass `-T` to list out the callable tasks
(like cap/rake).

In a future effort, some of the UX will be polished but I want
to wait and test in real-world scenarios with real-world tasks.

A few notes:

* I added the `reset` API to the main module - this is mostly
  to help with testing the CLI
* The Config now raises an exception when trying to lookup an
  unknown task - this is so the CLI can rescue this exception
  and handle it in a nice way
* The Config `task` DSL method now just returns the named task
  if no task class is specified.  The CLI uses this method to
  lookup the tasks.
* The TreeRunner now expects the Kernel to be passed in on init
  so it can `puts` the task tree output.  I need to use `puts`
  and not the logger b/c the logger has been disabled in the tree
  runner.