add testing helper methods for Task ssh hosts and callbacks
These allow the user to test without having to know details about
how the task implements its internal logic.  This adds class
methods for easily getting the callback task classes that were
configured.  This also makes the `dk_dsl_ssh_hosts` method public
so you can use it to test the ssh hosts that were configured.
This needs to be an instance method as you can configure procs
that are eval'd against the task instance and use instance data
to determine the hosts.

This all came up while testing tasks in a coming Dk::ABDeploy
gem.