(hotfix) force ssh hosts values to an Array
This should have been done originally but was missed b/c it only
causes an error in ruby 2.3.  String doesn't respond to `sort` in
2.3 but does in 1.8.7.  So if you pass it a string value for hosts
it would try and sort the hosts and fail in 2.3.

I want to support sending a string value for hosts.  Not only is
it a reasonable thing if just wanting a single host, in tests you
may set a host group but not actually configure that group.  In
this case, the group name is assumed to be the host name.

This came up while testing dk-abdeploy in 2.3.