add `ssh_cmd_str` Task helper method
This method is to help running local ssh cmds against a single
host.  It uses all the same remote ssh cmd logic to build the
local cmd string and lookup/handle any ssh args.

To accomplish this I added a `ssh_cmd_str` singleton method to
`Remote`.  This commonizes the ssh cmd building logic and allows
me to call for the ssh cmd str from a task instance.  The task
instance uses the same logic as the `ssh` helpers to lookup ssh
args so all ssh configs are honored.

Note: this cleans up the tests a bit as I noticed some things
while writing the tests for this effort.