(hotfix) add `ssh_cmd_str` test helper
This is designed to help test local cmds the run ssh cmd strings.
This keeps the user from knowing the internals about how the
`ssh_cmd_str` method works and allows them to just test their own
logic and not the internals.

This came up when testing a task that makes a ssh cmd str local
cmd.