change Config to store task callbacks
This changes the Config class to actual store callback config info
instead of blindly calling the Task DSL.  The idea is that config
settings should be confined to just the config instance.  The prev
implementation would take config settings and apply them to the
task singletons (changing global data).  This is not friendly as
the user is making instance changes that are actually making
global changes.  If the user intends to add a global callback,
they can do that via the normal task DSL api.

This switches to storing the settings on the config instance and
then making the task aware of these when running its callbacks.
I chose to store 4 callback collections on the config b/c I need
to differentiate between normal and prepend callbacks so that the
prepend can be applied before the regular callbacks and the regular
global callbacks.

Note: outside of the global task callbacks, this has no effect
to the overall behavior.