add Task DSL method to ensure tasks run only once
This is useful for "gatekeeper" tasks. For example, many different
tasks could setup a "Validate" task as a before callback.  However
the task only needs to run once, even if multiple tasks that are
being run set it up as a before callback.

This updates the runner to track a Set of task classes that have
run.  The task now checks whether it is configured to only run
once and, if so, halts if it has already run.  All callbacks are
always run.