Pass stdin to local and remote cmds
This updates the `Task` and `Runner` to allow passing a stdin
value when running local and remote commands. This allows using
stdin without having to embed it into a cmd str.

This updates the `Task` to allow optionally passing a stdin value
to the `cmd` and `ssh` helper methods. The helpers now take a
cmd str, input and opts. The input and opts are both optional and
either one can be provided without the other. This keeps the
interface backwards compatible when running commands without a
stdin value.

This also updates the `Runner` and `TestRunner` to expect a cmd
input value. This is passed to the `Local` or `Remote` commands
`run` methods. The commands were already setup to allow passing
stdin values so this just adds the necessary plumbing.

Finally, this updates the `Local` and `Remote` cmd spies to provide
a helper for testing the stdin value they were run with. This
allows testing that commands are run with an expected stdin.