Fix `cmd!` wrong number of args error
This fixes an error with `cmd!` calling `Runner#cmd` with the wrong
number of args. This is a bug from PR #39 that was missed because
the previous tests didn't test calling `cmd!` with all the possible
args variations. This fixes `cmd!` by switching it to call
`Task#cmd` instead of directly calling `Runner#cmd`. This makes
`cmd!` use the logic to handle optional input and opts args before
`Runner#cmd` is called.

This also updates the task tests to call both `cmd!` and `ssh!`
with all the variations of args that they support. This caused the
previous code to fail as expected and provides us better coverage
going forward.

Finally, this also includes a minor tweak to the
`Remote.ssh_cmd_str` method. If a host isn't provided it won't
try to lookup host ssh args. This was causing strange behavior
that would modify a runner's host ssh args adding an empty string
`""` key that used the default ssh args (also an empty string).
This was because the host ssh args default automatically takes any
key it is passed and sets its value to the default ssh args. This
avoids that happening by chance when a host is not provided for
an ssh command.