Rework ssh stubbing to avoid having to know ssh opts
This reworks the ssh stubbing to avoid having to know the ssh opts
that are built by dk. This makes it so stubbing an ssh call matches
how the `ssh` method is called which greatly simplifies its usage
and avoids having to know how dk builds ssh opts using its
current configuration.

This changes the `ssh` method on `Runner` to now take given opts
and ssh opts. Previously it only took ssh opts. The base `Runner`
only uses the ssh opts as it did previously but by taking the
given opts as well this allows the `TestRunner` to use it for
stubbing. Similarly, the `Runner` how expects input to be passed
when building cmds. The input is not used by the base `Runner` but
will be used by the `TestRunner`. The base `Runner` has to be
updated to keep its interface consistent with what the `TestRunner`
needs.

Now that the `TestRunner` is passed the input, given opts and ssh
opts when building cmds they can be used with the stubbing. Stubs
are now keyed using the cmd str, input and given opts. Using the
given opts allows the opts that are used to call `ssh` can be used
when stubbing `ssh`. When building the cmd spy though the cmd str
and ssh opts are used as they previously were. This also changes
the stub methods to use the same interface as the `ssh` and `cmd`
methods where they will take both an input and opts but both are
optional. All of these changes allow stubbing cmds with the same
interface that they are called.

This also modifies the `TestRunner` to lazily eval its blocks.
Now they aren't called until the `cmd` or `ssh` methods are called
when a task is run. This doesn't change a lot of the behavior
except that the block is not run immediately so its logic is not
immediately available. This is more consistent with assert stubbing
though so it provides a more consistent experience.

Finally, this updates the `cmd` and `ssh` methods on `Task` to
default their opts the same way as the stubs. This fixes a bug
where stubs were not correctly matching because they stubs used
`nil` for opts when none were passed but the `Task` was using
`{}`. This lines them both up to use `nil` as the default if no
opts are passed which ensures stubs match as expected.