(hotfix) fix tests for ruby2; make Remote complain if given `nil` hosts
The Task private method `dk_lookup_ssh_hosts`, if no hosts
settings are given or defaulted, will return `[nil]` in ruby 1.8.7
but will return `[]` in ruby 2.  This was causing task tests to
pass in ruby 1.8.7 that should have been failing b/c no hosts were
configured or given.  The test were failing in Ruby 2 as expected
but we had missed this b/c we forgot to test in ruby 2.

This updates the task tests to configure default hosts for the
tasks so they won't fail in ruby 2 and won't eval hosts to `[nil]`
in 1.8.7.  This gets the Task tests working as expected.

This also updates the Remote host validation logic to look for
`nil` items in any given host lists.  This is not only a more
robust check that can reveal unintended errors, it also would have
revealed the bad Task tests and made them fail in 1.8.7.