add `--verbose` CLI flag
This flag set the stdout log level to 'debug' instead fo the
default 'info' level.  To do this feature, I had to build into
the Config the ability to set a stdout log level and then honor
it in the config's logger.  I also added the option to the CLI
and had it set the 'debug' log level if the option is present.