add `:dry_tree_run` cmd/ssh opt to run cmds w/ the dry/tree runners
Some tasks may rely on the output of a cmd or some state set by
an ssh to proceed and if those cmds aren't run the task will fail.
The dry/tree runners by default never run cmds (they just spy
them).  For these cmds that are essential to the task, use this
new option to force that they always run so you can still take
advantage of the `--dry-run` and `--tree` CLI options.