have the TreeRunner output the tasks that were run
This updates the TreeRunner so that it outputs the tasks that
were run in a tree display.  It counts the run number associated
with each top-level task that is run and for each outputs a tree
display of the task and any sub-tasks (recusively) that were run
b/c of the top-level task.

```
$ dk --tree test-task test-task
building task trees...
 1) Dk::TreeRunner::TestTask
    |-- Dk::TreeRunner::TestTask::SubTask
        |-- Dk::TreeRunner::TestTask::SubSubTask
 2) Dk::TreeRunner::TestTask
    |-- Dk::TreeRunner::TestTask::SubTask
        |-- Dk::TreeRunner::TestTask::SubSubTask

```

Note: the run numbers in the output have been justified to
support running up to 99 top-level tasks at once without any
formatting issues.

The goal here is to help debug/validate your tasks and make sure
what will be run is what you expect to be run.