make the `run_only_once` task behavior more robust
Before the run once logic was only tested at the Task's `run!`
method.  While this prevented any logic from running multiple
times, it had the overhead of the task's callbacks continuing to
be run even though the task itself and any tasks run in the `run!`
method would not be run.  This was inconsistent and not in the
spirit of the feature.

I noticed this when debugging the `--tree` CLI option.  I noticed
I had repeated tree entries for run-only-once tasks.

This moves the checking logic to the runner level.  Now if a
task is run-only-once, it is just built and returned and never
run at all.  Otherwise it is built and run normally.  This
prevents the TreeRunner from ever building and logging a task
run for the task and fixes the `--tree` CLI option issue.