track task callbacks using a CallbackSet
This is an array that behaves like a Set.  It prevents duplicate
callbacks from being added to any one callback group.  Also, when
the Config callbacks are combined with the Task callbacks, a set
is used to prevent any duplicates across groups.  The only
duplicates that will run are if one callback is added as both a
"before" callback and an "after" callback.

The goal of all this is to be forgiving in the case that mixins
or 3rd-party extensions unknowingly add the same callback twice.
It also helps in the case a task at run-time adds a callback. We
don't want to duplicate callbacks each time a task is run.

I chose to implement a custom set object both b/c I added a custom
method to combine sets but also b/c Sets don't respond to the
push ('shovel', `<<`) and unshift methods.