Add `Dk::HasTheStubs` mixin, use in test and dry runner
This pulls the stubbing logic out of the `TestRunner` and adds it
to the mixin `HasTheStubs`. The mixin is added to both the
`TestRunner` and the `DryRunner` so that they can both use stubs.
This is setup for adding dry/tree stubs to the config and using
them when a running tasks using a dry or tree runner.

This caused the `dry_tree_run` opt to be reworked because using
stubbing logic from the `TestRunner` caused it to no longer run
the cmds. The stubbing logic was modified to allow the runners to
define how `HasTheStubs` should build a cmd when the cmd isn't
stubbed. For the test runner it builds a spy for the cmd anyway
because we never want to run a cmd with the test runner. For the
dry and tree runner they will check if the `dry_tree_run` opt is
used. If so, they will build a real cmd/ssh otherwise they will use
a spy. This removes all of the old logic for handling
`dry_tree_run` and also removes setting scmd's test mode env var as
its no longer needed to keep the dry/tree runner from running cmds.

Removing setting scmd's test mode env var revealed an issue with
using scmd's `CommandSpy`. It previously didn't work if the env
var wasn't set. This updates to the latest version of scmd so that
the `CommandSpy` can be used without having to set the env var.

This also moves the stub tests from the `TestRunner` tests and
reworks them to be system tests for `HasTheStubs`. This ensures
that they work with the `TestRunner`, `DryRunner` and `TreeRunner`.
This also expands them to ensure that they work with all possible
combinations of args (cmd str only, cmd str and input, cmd str and
opts or all of them).

This also adds unit tests for the `HasTheStubs` mixin. These test
the public stub cmds but also test the private `build_local_cmd`
and `build_remote_cmd` methods. These are tested even though they
are private because they are the methods that integrate with the
runners. This ensures they work as expected and that they should
integrate with the runners as expected.

Finally, this adds system tests for the `dry_tree_run` logic to
examine how it behaves with all of the different runners. This
ensures that we don't accidentally lose its behavior with future
changes.