Add storing dry/tree runner stubs to `Config`
This updates the `Config` to allow storing dry/tree runner stubs.
These will be used to add stubs to the runners when they are built.
This will allow a cmds output or exitstatus to be changed from the
default when a dry and tree runner runs. Specifically this is
needed when a task uses a cmds output or exitstatus to change the
flow of a task.

This adds DSL methods to the `Config` for adding cmd and ssh stubs.
These are simply stored on the `Config` and in a future effort they
will be applied to a dry/tree runner. This will allow a dk gem or
app using dk to include stubs for cmds to ensure they can be used
with a dry or tree runner.