have tasks log when they run
This includes identifying the task class that is starting to run
and identifying when a task class finishes running (including the
formatted run time of the task).  This gives the user timing info
on their tasks and let's them know when any one task starts or
finishes so they are aware of the progress.

This is part of cleaning up and making the both the file and
stdout UX more friendly.

```
$ dk my-task
> MyTask ...
> MySubTask ...
  ... MySubTask (572.1ms)
  ... MyTask (0:02s)
$
```

Note: this also updates to the latest Assert which has a float
Factory bugfix needed for these tests.