Update `DryRunner`, apply dry tree stubs from `Config` on init
This updates the `DryRunner` to apply dry tree stubs from it's
`Config` when it is initialized. This is the last part of allowing
users to add dry/tree runner stubs when configuring dk so they can
use the `--dry-run` and `--tree` options when running a task that
uses cmds output or exitstatus to change its flow.

The `DryRunner` now looks at its config's dry tree stubs and for
each of them it adds it's own stubs that match. This makes it so
a dry or tree runner will always use any stubs that are configured
on dk.

This also updates `HasTheStubs` to provide a `local_cmd_stubs` and
`remote_cmd_stubs` methods for testing and debugging. These can be
used to easily check the stubs that are configured on a runner.
This was noticed while implementing the `DryRunner` unit tests and
realizing it would require checking private methods and knowing
a lot of details about how `HasTheStubs` stores stubs.

Finally, this renames a few methods in `HasTheStubs`. The hash that
stores the blocks from the stub methods was renamed from "spy
blocks" to "stub blocks" since they come from the stub methods and
represent what cmds have been stubbed. This is more consistent with
the new public method that return `Stub` structs (we build `Stub`
structs from the stub blocks). The `cmd_spy_key` method was also
renamed to `has_the_stubs_key`. This is a more generic name since
it is used to key both the cmd spies and the stub blocks.

Closes #43