Log a task only around its run, not around its callbacks
This changes the logging of a task to only be around its run and
not around its callbacks. This seperates the tasks in the logs and
makes them easier to read.

This moves the before and after callbacks out of the block that is
passed when logging the task run. This means the callbacks will be
logged separate from the original task.

This also updates the cli logging to add a footer and time the
entire running of the cli call. This partially replaces the
previous behavior in that it can give you timing information for
how long a task and all of its callbacks took to run. This also
is a new feature in the case multiple tasks are run since it shows
the total time it took to run all of the tasks. Finally, the footer
provides a nice bookend when looking at a log file. This makes it
even easier to see where one cli run ends and another begins.

Closes #58