pre-check all task names before running the tasks
The goal here is more friendly UX.  There is no need to start
running tasks if one of the named tasks the CLI is given is
unknown/invalid.  The switches to pre-checking the given task
names.

This pre-check raises the same UnknownTaskError the config does
on task lookup.  This means the same error handling will be used.
This does tweak that error handling to put the error messages
below the help output so the user will see them in their terminal.

```
$ dk whatever not-known
...

Options:
    -T, --[no-]list-tasks            list all tasks available to run
    -d, --[no-]dry-run               run the tasks without executing any local/remote cmds
    -t, --[no-]tree                  print out the tree of tasks/sub-tasks that would be run
    -v, --[no-]verbose               run tasks showing verbose (ie debug log level) details
        --version
        --help

No task named "whatever", "not-known"
$
```